Precio arreglado

+++ backend/applications/Ventas/serializers/serializers.py
@@ -15,7 +15,7 @@ class CrearPedidoSerializer(serializers.Serializer):
     metodo_pago = serializers.ChoiceField(
         choices=[(role.value, role.name) for role in MetodoPagoEnum],
     )
-    precio = serializers.IntegerField()
+    precio = serializers.FloatField()
     metodo_entrega = serializers.ChoiceField(
         choices=[(role.value, role.name) for role in MetodoEntregaEnum]
     )
